Add "hex key add <keyname>" command
The "add" command adds a new API key to the user's Hex account. The key is
displayed, but not stored anywhere.

Also includes some refactoring of the rebar3_hex_user and rebar3_hex_key
modules so that the call each other. Also use the current username (if
present) when prompting for username/password.