Get correct password on win32.
This changes the password routines on windows to ignore the newline on password entry.

In addition, this adds additional overwriting on the clear routines to both hide the ANSI escape code when it is ignored and to overprint passwords less than 16 characters with spaces.

There's still some edge cases with narrow terminals, but this allows for authentication and publishing to work on windows again.