Handle all types of deps in has_unstable_deps check (#223)
When the has_unstable_deps check was introduced non-hex based deps were not taken into account, this commit
fixes this. What this fix doesn't address is whether git deps should be considered unstable, which is a moot point
since we don't allow publishing when the project has a git dep and will error out anyway.