Add V7 publish options
This commit constitutes the breaking interface changes for the publish command for V7.

Options aside, one of the biggest changes is removing umbrella related functionality. Specifically, instead of looking
for all possible apps and iterating over them and optionally publishing them, we just get the current app and publish it.

In addition, rebar3_hex:command_state/1 has been added, it's not extensively used yet. We will expand on this function,
what it returns exactly, and proliferate its usage in the publish command and all other commands in subsequent
commits.

 - add package task on publish command
 - add docs task support on publish command
 - remove --without-docs option on publish command
 - remove --package option that goes with --revert on publish command
   - This was only ever needed per umbrella support
 - Remove umbrella support (i.e., we only publish the current app vs all possible apps)
 - add command_state/1 on rebar3_hex