Don't require repo arg for building under certain conditions
Most cases for using the build provider shouldn't require specifying a repo argument.
The exception is when there is more than one repo found and at least one has doc config set on the
repo.

In addition, configuring documentation providers within repo config is deprecated and a warning will
be emitted if such config is detected.